requirements.txt
anthropic>=0.40.0
yfinance>=0.2.50
alpaca-py>=0.30.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pyyaml>=6.0
pytz>=2024.1
